requirements.txt
absl-py==0.13.0
#aiobotocore==1.3.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
argon2-cffi==20.1.0
arrow==1.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
binaryornot==0.4.4
bleach==3.2.3
#boto3==1.17.79
#botocore==1.20.49
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.0.0
cloudpickle==1.6.0
colorama==0.4.4
cookiecutter==1.7.3
cycler==0.10.0
dask==2021.6.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.1.0
flatbuffers==1.12
fsspec==2021.5.0
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
partd==1.2.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
poyo==0.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2020.5
#pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.0
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
s3fs==2021.5.0
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
statsmodels==0.12.1
tabulate==0.8.7
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
xlrd==2.0.1
XlsxWriter==1.3.7
yarl==1.6.3
zipp==3.4.0
zope.interface==5.2.0
gunicorn
flask
flask-restful
flask_cors
pyopenssl
gzip-reader